requirements.txt
fastapi==0.116.1
pydantic==2.11.7
uvicorn==0.35.0
streamlit==1.48.0
pandas==2.3.1
plotly==6.2.0
pytest==8.4.1
requests==2.32.4
python-dotenv==1.1.1
